% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLevelNavbar.R
\name{dropDownList}
\alias{dropDownList}
\title{Creates a drop-down menu}
\usage{
dropDownList(title, ...)
}
\arguments{
\item{...}{any number of menu items or dropdowns}

\item{the}{label of the drop down}
}
\description{
Creates a drop-down menu
}

