% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLevelNavbar.R
\name{lineItem}
\alias{lineItem}
\title{Creates a menu action item}
\usage{
lineItem(title, value = title)
}
\arguments{
\item{title}{the displayed label for the menu item}

\item{value}{return value upon clicking}
}
\description{
Creates a menu action item
}

