% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLevelNavbar.R
\name{multiLevelNavBarPage}
\alias{multiLevelNavBarPage}
\title{Creates the top level menu bar}
\usage{
multiLevelNavBarPage(..., title = "", id = NULL, theme = NULL)
}
\arguments{
\item{...}{any number of menu items or dropdowns}

\item{title}{(optional) the title of the menubar}

\item{id}{the id to be associated with this menubar}

\item{theme}{the name of a shiny bootstrap theme. (requires shinythemes package)}
}
\description{
Creates the top level menu bar
}

