% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMultilevelMenu.R
\name{updateMultiLevelMenu}
\alias{updateMultiLevelMenu}
\title{Update the Menu}
\usage{
updateMultiLevelMenu(session, menuBarId, command, target, type, ...)
}
\arguments{
\item{session}{the session}

\item{menuBarId}{the id of the menubar to be updated}

\item{command}{one of c('disable','enable', 'rename')}

\item{target}{the identifier(value) of dropdown/selection}

\item{type}{one of c('actionItem', 'dropdown')}

\item{...}{additional params}
}
\description{
Update the Menu
}

