% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLevelNavbar.R
\name{multiLevelNavBarPage}
\alias{multiLevelNavBarPage}
\title{Creates the top level menu bar}
\usage{
multiLevelNavBarPage(..., navBarBrand = "", id = NULL)
}
\arguments{
\item{...}{any number of menu items or dropdowns}

\item{navBarBrand}{(optional)}

\item{id}{the id to be associated with this menubar}
}
\description{
Creates the top level menu bar
}

