% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMultilevelMenu.R
\name{updateMultiLevelMenu}
\alias{updateMultiLevelMenu}
\title{Update the Menu}
\usage{
updateMultiLevelMenu(session, menuBarId, command, targetItem, type, ...)
}
\arguments{
\item{session}{the session}

\item{menuBarId}{the id of the menubar to be updated}

\item{command}{one of c('disable','enable', 'rename')}

\item{targetItem}{the identifier(value) of dropDown/selection}
}
\description{
Update the Menu
}

